Allow doubles in data points

+++ verifier/Verifier/SignatureDataPoint.cs
@@ -2,11 +2,11 @@ namespace Verifier;
 
 public record struct SignatureDataPoint(
     long TimeStamp,
-    int XCoord,
-    int YCoord,
-    int Pressure,
-    int AltitudeAngle,
-    int AzimuthAngle,
-    int Height,
-    int Twist
+    double XCoord,
+    double YCoord,
+    double Pressure,
+    double AltitudeAngle,
+    double AzimuthAngle,
+    double Height,
+    double Twist
 );
